feat: 3d point plots
chore: annotate axis labels, file relocation

chore: refactoring, separate plot functions from analysis
