+++ perception_eval/perception_eval/tool/__init__.py
@@ -4,7 +4,7 @@ from .gmm import Gmm
 from .gmm import load_sample
 from .perception_analyzer2d import PerceptionAnalyzer2D
 from .perception_analyzer3d import PerceptionAnalyzer3D
-from .perception_analyzer3d_field import PerceptionAnalyzer3DField, PerceptionFieldXY, PerceptionFieldAxis, DataTableIdx
+from .perception_analyzer3dfield import PerceptionAnalyzer3DField, PerceptionFieldXY, PerceptionFieldAxis, DataTableIdx
 from .utils import PlotAxes
 
 # type aliases
