+++ perception_eval/perception_eval/visualization/__init__.py
@@ -16,6 +16,7 @@ from typing import Union
 
 from perception_eval.visualization.perception_visualizer2d import PerceptionVisualizer2D
 from perception_eval.visualization.perception_visualizer3d import PerceptionVisualizer3D
+from perception_eval.visualization.perception_visualizer3dfield import PerceptionFieldPlot, PerceptionFieldPlots
 from perception_eval.visualization.sensing_visualizer import SensingVisualizer
 
 # type aliases
